Fixes false positives!
First you get false negatives, then you get false positives. This makes
sure that an array is always returned from the testing methods. This
makes figuring out whether the test passed or failed much easier.
Basically, if the returned array contains a value of "false" the test
fails. Prior, we were doing type checking and reducing arrays down to
single values and it was a messsssss.